rename: 지원한 팟, 내가만든 팟 페이지 파일 이름 수정
