+++ src/pages/Pots/AppliedPot/AppliedPot.style.ts
