+++ src/pages/Pots/MadePot/MadePot.style.ts
