+++ src/pages/index.ts
@@ -8,8 +8,8 @@ import SearchResultPage from "./SearchResult/SearchResult";
 import PotDetailPage from "./PotDetail/PotDetail";
 import PotMain from "./Pots/PotMain";
 import AllPotPage from "./Pots/AllPot/AllPotPage";
-import AppliedPotPage from "./Pots/AppliedPot/AppliedPotPage";
-import MadePotPage from "./Pots/MadePot/MadePotPage";
+import AppliedPotPage from "./Pots/AppliedPot/AppliedPot";
+import MadePotPage from "./Pots/MadePot/MadePot";
 import SignUpPage from "./SignUp/SignUp";
 import CreatePotPage from "./CreatePot/CreatePot";
 import WritingPage from "./WritePost/WritePost";
